Support decimal movement costs throughout + terrain symbol graphics
- config.py: switch terrain_base_costs, combination_costs,
  get_move_cost and river_movement_penalty from int to float so
  values like 2.5 / 3.5 / 1.5 in terrain.config are preserved
- game_state.py: Tile.move_cost, Player.movement_remaining/total and
  MoveResult/MoveToResult cost fields changed to float
- game.py: Dijkstra dist/heap types updated to float, round() removed
  from diagonal step cost, messages formatted with :g (strips .0)
- terrain.config: combination costs retuned with decimal increments
  (hills = base+2.0, mountain = base+3.5) for smoother progression;
  pre-existing integer rounding was masking fractional base costs
- app.js: add drawTerrainSymbol() — hill tiles get two white semi-
  circle bumps, mountain tiles get a triangle with a snow cap
  overlay; symbols scale with zoom and render in terrain map mode

+++ src/landgrab/models/game_state.py
@@ -47,7 +47,7 @@ class Tile(BaseModel):
     modifier: ModifierType = ModifierType.FLAT
     elevation: float = Field(ge=0.0, le=1.0)
     hills_scalar: float = Field(default=0.0, ge=0.0, le=1.0)
-    move_cost: int | None = None  # None = impassable
+    move_cost: float | None = None  # None = impassable
     is_river: bool = False
 
 
@@ -63,8 +63,8 @@ class Player(BaseModel):
     name: str
     position: tuple[int, int]
     gold: int = 100
-    movement_remaining: int = 0
-    movement_total: int = 0
+    movement_remaining: float = 0
+    movement_total: float = 0
 
 
 class GameState(BaseModel):
@@ -109,8 +109,8 @@ class MoveResult(BaseModel):
     new_position: tuple[int, int]
     terrain: TerrainType
     modifier: ModifierType
-    move_cost: int
-    movement_remaining: int
+    move_cost: float
+    movement_remaining: float
     message: str
     game_state: GameState
 
@@ -119,8 +119,8 @@ class MoveToResult(BaseModel):
     new_position: tuple[int, int]
     terrain: TerrainType
     modifier: ModifierType
-    total_cost: int
-    movement_remaining: int
+    total_cost: float
+    movement_remaining: float
     path: list[tuple[int, int]]
     message: str
     game_state: GameState
